Use open_memstream(3) to work around broken macOS buffered output
Passing NULL as the fp argument of dtrace_work results in segfaults on
macOS. Work around this by passing a FILE * obtained from open_memstream
and reading it after each call to dtrace_work().
This should generally result in the same output (the hello world test now
passes on macOS), but there could be some corner cases that are different.